requirements.txt
numpy
pandas
scikit-learn
scipy
joblib
requests
python-dotenv
psycopg2-binary